+++ core/CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add new protocol version 1.18.
 - Add `SubscribeService`, `SubscribeServiceReply` and `UnsubscribeService` messages.
-- Add `SubscribeAllEvents` and `SubscribeAllEventsReply` messages.
+- Add `SubscribeAllEvents`, `SubscribeAllEventsReply` and `UnsubscribeAllEvents` messages.
 
 ## [0.7.0] - 2024-07-25